+++ gatsby-node.js
@@ -3,4 +3,4 @@
  */
 module.exports.modifyWebpackConfig = require("./gatsby/modifyWebpackConfig");
 module.exports.createPages = require("./gatsby/createPages");
-module.exports.onCreatePage = require("./gatsby/onCreatePage")
+module.exports.onCreatePage = require("./gatsby/onCreatePage");
