+++ src/Auth/authInstance.js
@@ -1,6 +1,6 @@
 import buildConfig from "../../build.config.json";
-import ailoAuth0 from "./ailo_Auth";
-import irisAuth from "./iris_Auth";
+import ailoAuth0 from "./ailo_auth";
+import irisAuth from "./iris_auth";
 
 export default (process.env.NODE_ENV === "development"
   ? ailoAuth0
