+++ src/NewComponent/Header/Header.js
@@ -2,7 +2,7 @@ import React from "react";
 import Link from "gatsby-link";
 import PropTypes from "prop-types";
 import {CSSTransition} from "react-transition-group";
-import IRISAuth from "../../Auth/IRISAuth";
+import Auth from "../../Auth";
 
 import imgLogo from "../../assets/images/NewDesign/Header/logo.svg";
 import icClose from "../../assets/images/NewDesign/Header/ic-close.svg";
@@ -15,6 +15,7 @@ const secondaryIndexes = [2, 6, 10, 14];
 const primaryIndexes = [0, 1, 3, 4, 5, 7, 8, 9, 11, 12, 13, 15, 16, 17, 18];
 var menuClass = "";
 var btnClass = "";
+
 function getLoginLogout(auth) {
   const {auth: authVar, env} = config;
   const {login, logout, isAuthenticated} = auth;
@@ -74,7 +75,7 @@ class Header extends React.Component {
       menuClass = "secondary";
     }
     return (
-      <IRISAuth
+      <Auth
         render={auth => (
           <nav className={`navbar is-transparent header-wrapper`}>
             <div className="navbar-brand">
