+++ src/pages/foundry/index.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 import Link from "gatsby-link";
 import Img from "gatsby-image";
 import YouTube from "react-youtube";
-import IRISAuth from "../../Auth/IRISAuth";
+import Auth from "../../Auth";
 import "./styles.scss";
 
 const videoOptions = {};
@@ -41,7 +41,7 @@ class MyFoundryPage extends React.Component {
       },
     } = this.props;
     return (
-      <IRISAuth
+      <Auth
         render={auth => {
           this.auth = auth;
           return (
